backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic[email]==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai>=0.4.0
pillow==10.2.0
langchain==0.1.6
langchain-google-genai==0.0.11
langchain-core==0.1.23
openai>=1.10.0
httpx
python-decorators
boto3==1.26.137